A few minor optimizations:

* a small optimization that avoids calling `len(slots)` twice.
* Extend the slots pool with a tuple of workers instead of a list.
  The performance difference is in nanoseconds but it took me just a second to make the change.

> In [4]: %timeit l=[];l.extend((foo,) * 6)
> 166 ns ± 3.72 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)
> 
> In [5]: %timeit l=[];l.extend([foo,] * 6)
> 191 ns ± 0.823 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)