@erdewit 2018-01-18
Thanks for the optimizations, looks good to me.